Create version 0.2
Added in a function that searches the coordinate space provided by the "generatecoord" function, and one to manage the search and saving functions. In effect, this version can "learn" a level. There is not playback function yet, however, nor is it capable of beating holes which take more than 1 shot. It should be *possible* for it to learn any hole it can hit a hole-in-one on.